Move shared component to components dir
