+++ frontend/src/pages/instance/components/InstanceLogs.vue
